refactor: Move rule & channel yamls into yaml domain
