+++ python/lib/sift_py/ingestion/config/yaml/spec.py
@@ -4,9 +4,19 @@ Formal specification of the types that `sift_py` expects when loading a telemetr
 
 from __future__ import annotations
 
-from typing import Dict, List, Literal, Union
+from typing import Dict, List
 
 from typing_extensions import NotRequired, TypedDict
+
+import sift_py.yaml.channel as channel_yaml
+import sift_py.yaml.rule as rule_yaml
+
+RuleYamlSpec = rule_yaml.RuleYamlSpec
+RuleNamespaceYamlSpec = rule_yaml.RuleNamespaceYamlSpec
+NamedExpressionYamlSpec = rule_yaml.NamedExpressionYamlSpec
+ChannelConfigYamlSpec = channel_yaml.ChannelConfigYamlSpec
+ChannelEnumTypeYamlSpec = channel_yaml.ChannelEnumTypeYamlSpec
+ChannelBitFieldElementYamlSpec = channel_yaml.ChannelBitFieldElementYamlSpec
 
 
 class TelemetryConfigYamlSpec(TypedDict):
@@ -29,58 +39,6 @@ class TelemetryConfigYamlSpec(TypedDict):
     flows: NotRequired[List[FlowYamlSpec]]
 
 
-class ChannelConfigYamlSpec(TypedDict):
-    """
-    Formal spec that defines what a channel should look like in YAML.
-
-    `name`: Name of channel.
-    `description`: Optional channel description.
-    `unit`: Unit of measurement.
-    `component`: Name of component that channel belongs to.
-    `data_type`: Type of the data associated with the channel.
-    `enum_types`: Required if `data_type` is `enum.
-    `bit_field_elements`: Required if `data_type` is `bit_field`.
-    """
-
-    name: str
-    description: NotRequired[str]
-    unit: NotRequired[str]
-    component: NotRequired[str]
-    data_type: Union[
-        Literal["double"],
-        Literal["string"],
-        Literal["enum"],
-        Literal["bit_field"],
-        Literal["bool"],
-        Literal["float"],
-        Literal["int32"],
-        Literal["int64"],
-        Literal["uint32"],
-        Literal["uint64"],
-    ]
-    enum_types: NotRequired[List[ChannelEnumTypeYamlSpec]]
-    bit_field_elements: NotRequired[List[ChannelBitFieldElementYamlSpec]]
-
-
-class ChannelEnumTypeYamlSpec(TypedDict):
-    """
-    Formal spec that defines what a channel enum type should look like in YAML.
-    """
-
-    name: str
-    key: int
-
-
-class ChannelBitFieldElementYamlSpec(TypedDict):
-    """
-    Formal spec that defines what a bit-field element should look like in YAML.
-    """
-
-    name: str
-    index: int
-    bit_count: int
-
-
 class FlowYamlSpec(TypedDict):
     """
     Formal spec that defines what a flow should look like in YAML.
@@ -88,124 +46,6 @@ class FlowYamlSpec(TypedDict):
 
     name: str
     channels: List[ChannelConfigYamlSpec]
-
-
-class RuleNamespaceYamlSpec(TypedDict):
-    """
-    The formal definition of what a rule namespace looks like in YAML.
-
-    `namespace`: Name of the namespace.
-    `rules`: A list of rules that belong to the namespace.
-    """
-
-    namespace: str
-    rules: List[RuleYamlSpec]
-
-
-class RuleYamlSpec(TypedDict):
-    """
-    The formal definition of what a single rule looks like in YAML.
-
-    `name`: Name of the rule.
-    `namespace`: Optional namespace of the rule. Only used if referencing a rule defined in a namespace.
-    `rule_client_key`: User-defined string-key that uniquely identifies this rule config.
-    `description`: Description of rule.
-    `expression`:
-        Either an expression-string or a `sift_py.ingestion.config.yaml.spec.NamedExpressionYamlSpec` referencing a named expression.
-    `type`: Determines the action to perform if a rule gets evaluated to true.
-    `assignee`: If `type` is `review`, determines who to notify. Expects an email.
-    `tags`: Tags to associate with the rule.
-    `channel_references`: A list of channel references that maps to an actual channel. More below.
-    `sub_expressions`: A list of sub-expressions which is a mapping of place-holders to sub-expressions. Only used if using named expressions.
-    `asset_names`: A list of asset names that this rule should be applied to. ONLY VALID if defining rules outside of a telemetry config.
-    `tag_names`: A list of tag names that this rule should be applied to. ONLY VALID if defining rules outside of a telemetry config.
-
-    Namespaces:
-    Rule may be defined in a separate YAML within a namespace. The reference to the namespace rule would look like the following:
-    ```yaml
-    rules:
-      - namespace: voltage
-        name: overvoltage
-        channel_references:
-          - $1: *vehicle_state_channel
-          - $2: *voltage_channel
-    ```
-    With the corresponding rule being defined in a separate YAML file like the following:
-    ```yaml
-    namespace: voltage
-    rules:
-      - name: overvoltage
-        description: Checks for overvoltage while accelerating
-        expression: $1 == "Accelerating" && $2 > 80
-        type: review
-    ```
-
-    Channel references:
-    A channel reference is a string containing a numerical value prefixed with "$". Examples include "$1", "$2", "$11", and so on.
-    The channel reference is mapped to an actual channel config. In YAML it would look something like this:
-
-    ```yaml
-    channel_references:
-      - $1: *vehicle_state_channel
-      - $2: *voltage_channel
-    ```
-
-    Sub-expressions:
-    A sub-expression is made up of two components: A reference and the actual sub-expression. The sub-expression reference is
-    a string with a "$" prepended to another string comprised of characters in the following character set: `[a-zA-Z0-9_]`.
-    This reference should be mapped to the actual sub-expression. For example, say you have kinematic equations in `kinematics.yml`,
-    and the equation you're interested in using looks like the following:
-
-    ```yaml
-    kinetic_energy_gt:
-      0.5 * $mass * $1 * $1 > $threshold
-    ```
-
-    To properly use `kinetic_energy_gt` in your rule, it would look like the following:
-
-    ```yaml
-    rules:
-      - name: kinetic_energy
-        description: Tracks high energy output while in motion
-        type: review
-        assignee: bob@example.com
-        expression:
-          name: kinetic_energy_gt
-        channel_references:
-          - $1: *velocity_channel
-        sub_expressions:
-          - $mass: 10
-          - $threshold: 470
-        tags:
-            - nostromo
-    ```
-    """
-
-    name: str
-    namespace: NotRequired[str]
-    rule_client_key: NotRequired[str]
-    description: NotRequired[str]
-    expression: Union[str, NamedExpressionYamlSpec]
-    type: Union[Literal["phase"], Literal["review"]]
-    assignee: NotRequired[str]
-    tags: NotRequired[List[str]]
-    channel_references: NotRequired[List[Dict[str, ChannelConfigYamlSpec]]]
-    sub_expressions: NotRequired[List[Dict[str, str]]]
-    asset_names: NotRequired[List[str]]
-    tag_names: NotRequired[List[str]]
-
-
-class NamedExpressionYamlSpec(TypedDict):
-    """
-    A named expression. This class is the formal definition of what a named expression
-    should look like in YAML. The value of `name` may contain a mix of channel references
-    and channel identifiers.
-
-    For a formal definition of channel references and channel identifiers see the following:
-    `sift_py.ingestion.config.yaml.spec.RuleYamlSpec`.
-    """
-
-    name: str
 
 
 class YamlConfigError(Exception):
